fix: learn configuration flag to correctly encode index name in URL (#574)
We've recently identified that the user-provided index name, when passed
to the URL internally, was not URL-encoded as one would expect. This
could result in the user targetting wrong indices, either when using the Go
client, or other API clients, including frontend integrations.

Because URL-encoding this would be considered as a breaking-change, and
we cannot afford creating a new major version anytime soon, this commit
introduce a new field in the search.Client's configuration:

```go
search.Configuration{
  AppID:  "<your_app_id>",
  APIKey: "<your_api_key>",
  CompatibilityBreakingFeatures: search.CompatibilityBreakingFeatures{
    EnableURLEncodingOfIndexName: true,
  },
}
```

When `EnableURLEncodingOfIndexName` is absent (default behavior) or set
to false, the previously existing behavior is maintained. This prevents
any breaking-change on user side. However, by setting this paramater to
true, the index names will then be correctly encoded, effectively
producing the expected behavior.

Close #574